third_party/WebKit/LayoutTests/fast/css3-text/css3-text-justify/getComputedStyle/getComputedStyle-text-justify-expected.txt

This test checks that text-justify parses properly the properties from CSS 3 Text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Value 'none':
PASS element.style.textJustify is 'none'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'none'
PASS computedStyle.textJustify is 'none'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'none'

Value 'inter-word':
PASS element.style.textJustify is 'inter-word'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'inter-word'
PASS computedStyle.textJustify is 'inter-word'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'inter-word'

Value 'distribute':
PASS element.style.textJustify is 'distribute'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'distribute'
PASS computedStyle.textJustify is 'distribute'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'distribute'

Presetted value is none
Invalid value test - 'green':
PASS element.style.textJustify is 'none'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'none'
PASS computedStyle.textJustify is 'none'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'none'

Invalid value test - 'inline':
PASS element.style.textJustify is 'none'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'none'
PASS computedStyle.textJustify is 'none'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'none'

Presetted value is distribute
Invalid value test - 'solid':
PASS element.style.textJustify is 'distribute'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'distribute'
PASS computedStyle.textJustify is 'distribute'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'distribute'

Invalid value test - 'normal':
PASS element.style.textJustify is 'distribute'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'distribute'
PASS computedStyle.textJustify is 'distribute'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'distribute'

Computed value test - 'inherit':
PASS element.style.textJustify is 'inherit'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'inherit'
PASS computedStyle.textJustify is 'none'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'none'

Computed value test - 'inherit':
PASS element.style.textJustify is 'inherit'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'inherit'
PASS computedStyle.textJustify is 'distribute'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'distribute'

Computed value test - 'initial':
PASS element.style.textJustify is 'initial'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'initial'
PASS computedStyle.textJustify is 'auto'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'auto'

Value of ancestor is 'inter-word, while child is 'distribute':
PASS element.style.textJustify is 'distribute'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'distribute'
PASS computedStyle.textJustify is 'distribute'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'distribute'

Value of ancestor is 'none, while child is 'inter-word':
PASS element.style.textJustify is 'inter-word'
PASS element.style.getPropertyCSSValue('text-justify').cssText is 'inter-word'
PASS computedStyle.textJustify is 'inter-word'
PASS computedStyle.getPropertyCSSValue('text-justify').cssText is 'inter-word'

PASS successfullyParsed is true

TEST COMPLETE
